# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BhoomiScan Land Verification API
  version: 1.0.0
  description: |
    Backend API for land-record verification: mobile OTP login, Razorpay subscriptions,
    Surepass-backed land records (Punjab), historical snapshots with change detection,
    alerts, and smart-contract anchoring.

    **Envelope.** Success: `{ "success": true, "data": ..., "meta"?: ... }`.
    Error: `{ "success": false, "error": { "code", "message", "details"? }, "requestId" }`.

    **Auth.** `Authorization: Bearer <token>` from `POST /auth/verify-otp`.
servers:
  - url: http://localhost:4000/api/v1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          properties:
            code: { type: string, example: VALIDATION_FAILED }
            message: { type: string }
            details: {}
        requestId: { type: string }
...
    SmartContractRecord:
      type: object
      properties:
        id: { type: string, format: uuid }
        snapshotId: { type: string, format: uuid }
        provider: { type: string, enum: [HTTP, BSC, DISABLED] }
        status: { type: string, enum: [PENDING, PROCESSING, STORED, ALREADY_STORED, FAILED, PERMANENTLY_FAILED, SKIPPED] }
        dataHash: { type: string }
        reference: { type: string, nullable: true }
        transactionHash: { type: string, nullable: true }
        attempts: { type: integer }
        nextRetryAt: { type: string, format: date-time, nullable: true }
        errorCode: { type: string, nullable: true }
...
    Alert:
      type: object
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [LAND_RECORD_CHANGED, VERIFICATION_FAILED, SMART_CONTRACT_FAILED] }
        severity: { type: string, enum: [INFO, WARNING, CRITICAL] }
        title: { type: string }
        message: { type: string }
        data: { type: object }
        status: { type: string, enum: [UNREAD, READ] }
        createdAt: { type: string, format: date-time }
security:
  - bearerAuth: []
paths:
...
  /smart-contract/records:
    get:
      tags: [Smart Contract]
      summary: Anchoring records of a land verification
      parameters:
        - { name: landVerificationId, in: query, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      records: { type: array, items: { $ref: '#/components/schemas/SmartContractRecord' } }
  /smart-contract/records/{id}:
    get:
      tags: [Smart Contract]
      summary: One anchoring record
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { description: OK }
        '404': { $ref: '#/components/responses/Error' }
  /smart-contract/records/{id}/retry:
    post:
      tags: [Smart Contract]
      summary: Retry a FAILED / PERMANENTLY_FAILED record now
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { description: Result of the retry }
        '409': { $ref: '#/components/responses/Error' }
  /alerts:
    get:
      tags: [Alerts]
      summary: My alerts
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [UNREAD, READ] } }
        - { name: landVerificationId, in: query, schema: { type: string, format: uuid } }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      alerts: { type: array, items: { $ref: '#/components/schemas/Alert' } }
                      unread: { type: integer }
  /alerts/{id}/read:
    patch:
      tags: [Alerts]
      summary: Mark one alert read
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { description: OK }
...